sysusers: get rid of an unneccesary id line
This line is actually bad and will cause the creation of some users to
fail. The reason is simply because id -g arg can't take an argument of a
group that does not have a user. So it will fail here for many conf
files. This line doesn't really appear to have any particular purpose
either since add_user will check if ${id} is equal to "-" and handle
that logic for us. Just get rid of it altogether.